requirements.txt
Flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
Pillow
numpy
mysql-connector-python==8.1.0
tensorflow==2.13.0
kaggle==1.5.16
ultralytics==8.0.196
